[EST-41] Add: useMyLocation hook
